Move generic inlining context into own file
This is a refactoring and should not affect generated code.
