+++ source/slang/slang-ir-legalize-global-values.cpp
@@ -1,0 +1,236 @@
+#include "slang-ir-legalize-global-values.h"
+
+#include "slang-ir-clone.h"
+#include "slang-ir-util.h"
+
+namespace
+{
+using namespace Slang;
+void setInsertBeforeOutsideASM(
+    IRBuilder& builder, IRInst* beforeInst)
+{
+    auto parent = beforeInst->getParent();
+    while (parent)
+    {
+        if (as<IRSPIRVAsm>(parent))
+        {
+            builder.setInsertBefore(parent);
+            return;
+        }
+        parent = parent->getParent();
+    }
+    builder.setInsertBefore(beforeInst);
+}
+} // namespace
+
+namespace Slang
+{
+
+void GlobalInstInliningContextGeneric::inlineGlobalValues(IRModule * module)
+{
+    List<IRUse*> globalInstUsesToInline;
+
+    for (auto globalInst : module->getGlobalInsts())
+    {
+        if (isInlinableGlobalInst(globalInst))
+        {
+            for (auto use = globalInst->firstUse; use; use = use->nextUse)
+            {
+                if (getParentFunc(use->getUser()) != nullptr)
+                    globalInstUsesToInline.add(use);
+            }
+        }
+    }
+
+    for (auto use : globalInstUsesToInline)
+    {
+        auto user = use->getUser();
+        IRBuilder builder(user);
+        setInsertBeforeOutsideASM(builder, user);
+        IRCloneEnv cloneEnv;
+        auto val = maybeInlineGlobalValue(builder, use->getUser(), use->get(), cloneEnv);
+        if (val != use->get())
+            builder.replaceOperand(use, val);
+    }
+}
+
+bool GlobalInstInliningContextGeneric::isLegalGlobalInst(IRInst* inst)
+{
+    switch (inst->getOp())
+    {
+    case kIROp_MakeStruct:
+    case kIROp_MakeArray:
+    case kIROp_MakeArrayFromElement:
+    case kIROp_MakeVector:
+    case kIROp_MakeMatrix:
+    case kIROp_MakeMatrixFromScalar:
+    case kIROp_MakeVectorFromScalar:
+        return true;
+    default:
+        if (as<IRConstant>(inst))
+            return true;
+        if (isLegalGlobalInstForTarget(inst))
+            return true;
+        return false;
+    }
+}
+
+bool GlobalInstInliningContextGeneric::isInlinableGlobalInst(IRInst* inst)
+{
+    switch (inst->getOp())
+    {
+    case kIROp_Add:
+    case kIROp_Sub:
+    case kIROp_Mul:
+    case kIROp_FRem:
+    case kIROp_IRem:
+    case kIROp_Lsh:
+    case kIROp_Rsh:
+    case kIROp_And:
+    case kIROp_Or:
+    case kIROp_Not:
+    case kIROp_Neg:
+    case kIROp_Div:
+    case kIROp_FieldExtract:
+    case kIROp_FieldAddress:
+    case kIROp_GetElement:
+    case kIROp_GetElementPtr:
+    case kIROp_GetOffsetPtr:
+    case kIROp_UpdateElement:
+    case kIROp_MakeTuple:
+    case kIROp_GetTupleElement:
+    case kIROp_MakeStruct:
+    case kIROp_MakeArray:
+    case kIROp_MakeArrayFromElement:
+    case kIROp_MakeVector:
+    case kIROp_MakeMatrix:
+    case kIROp_MakeMatrixFromScalar:
+    case kIROp_MakeVectorFromScalar:
+    case kIROp_swizzle:
+    case kIROp_swizzleSet:
+    case kIROp_MatrixReshape:
+    case kIROp_MakeString:
+    case kIROp_MakeResultError:
+    case kIROp_MakeResultValue:
+    case kIROp_GetResultError:
+    case kIROp_GetResultValue:
+    case kIROp_CastFloatToInt:
+    case kIROp_CastIntToFloat:
+    case kIROp_CastIntToPtr:
+    case kIROp_PtrCast:
+    case kIROp_CastPtrToBool:
+    case kIROp_CastPtrToInt:
+    case kIROp_BitAnd:
+    case kIROp_BitNot:
+    case kIROp_BitOr:
+    case kIROp_BitXor:
+    case kIROp_BitCast:
+    case kIROp_IntCast:
+    case kIROp_FloatCast:
+    case kIROp_Greater:
+    case kIROp_Less:
+    case kIROp_Geq:
+    case kIROp_Leq:
+    case kIROp_Neq:
+    case kIROp_Eql:
+    case kIROp_Call:
+        return true;
+    default:
+        if (isInlinableGlobalInstForTarget(inst))
+            return true;
+        return false;
+    }
+}
+
+bool GlobalInstInliningContextGeneric::shouldInlineInstImpl(IRInst* inst)
+{
+    if (!isInlinableGlobalInst(inst))
+        return false;
+    if (isLegalGlobalInst(inst))
+    {
+        for (UInt i = 0; i < inst->getOperandCount(); i++)
+            if (shouldInlineInst(inst->getOperand(i)))
+                return true;
+        return false;
+    }
+    return true;
+}
+
+bool GlobalInstInliningContextGeneric::shouldInlineInst(IRInst* inst)
+{
+    bool result = false;
+    if (m_mapGlobalInstToShouldInline.tryGetValue(inst, result))
+        return result;
+    result = shouldInlineInstImpl(inst);
+    m_mapGlobalInstToShouldInline[inst] = result;
+    return result;
+}
+
+IRInst* GlobalInstInliningContextGeneric::inlineInst(IRBuilder& builder, IRCloneEnv& cloneEnv, IRInst* inst)
+{
+    IRInst* result;
+    if (cloneEnv.mapOldValToNew.tryGetValue(inst, result))
+        return result;
+
+    for (UInt i = 0; i < inst->getOperandCount(); i++)
+    {
+        auto operand = inst->getOperand(i);
+        IRBuilder operandBuilder(builder);
+        setInsertBeforeOutsideASM(operandBuilder, builder.getInsertLoc().getInst());
+        maybeInlineGlobalValue(operandBuilder, inst, operand, cloneEnv);
+    }
+    result = cloneInstAndOperands(&cloneEnv, &builder, inst);
+    cloneEnv.mapOldValToNew[inst] = result;
+    IRBuilder subBuilder(builder);
+    subBuilder.setInsertInto(result);
+    for (auto child : inst->getDecorations())
+    {
+        cloneInst(&cloneEnv, &subBuilder, child);
+    }
+    for (auto child : inst->getChildren())
+    {
+        inlineInst(subBuilder, cloneEnv, child);
+    }
+    return result;
+}
+
+IRInst* GlobalInstInliningContextGeneric::maybeInlineGlobalValue(
+    IRBuilder& builder,
+    IRInst* user,
+    IRInst* inst,
+    IRCloneEnv& cloneEnv)
+{
+    if (!shouldInlineInst(inst))
+    {
+        switch (inst->getOp())
+        {
+        case kIROp_Func:
+        case kIROp_Specialize:
+        case kIROp_Generic:
+        case kIROp_LookupWitness:
+            return inst;
+        }
+        if (as<IRType>(inst))
+            return inst;
+
+        // If we encounter a global value that shouldn't be inlined, e.g. a const literal,
+        // we should insert a GlobalValueRef() inst to wrap around it, so all the dependent
+        // uses can be pinned to the function body.
+        auto result = inst;
+        bool shouldWrapGlobalRef = true;
+        if (!isLegalGlobalInst(user) && !getIROpInfo(user->getOp()).isHoistable())
+            shouldWrapGlobalRef = false;
+        else if (shouldBeInlinedForTarget(user))
+            shouldWrapGlobalRef = false;
+        if (shouldWrapGlobalRef)
+            result = builder.emitGlobalValueRef(inst);
+        cloneEnv.mapOldValToNew[inst] = result;
+        return result;
+    }
+
+    // If the global value is inlinable, we make all its operands avaialble locally, and
+    // then copy it to the local scope.
+    return inlineInst(builder, cloneEnv, inst);
+}
+
+} // namespace Slang
